transfer House Name/Intro Fix module to House Manage Modal
