--- orange/src/components/House/HouseManageModal/HouseManageModal.css
@@ -1,14 +1,0 @@
-.house-toss-leader-form {
-  display: grid;
-  }
-
-.house-user-container {
-  width: 150%;
-  padding-bottom: 0.5em;
-}
-
-.member-box {
-  padding: 10px 15px 10px 15px;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
