+++ orange/src/components/House/HouseManageModal/HouseManageModal.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import React, { useCallback, useEffect, useState } from 'react';
+import EdiText from 'react-editext';
 import { AiOutlineCrown } from 'react-icons/ai';
 import { Select } from 'react-functional-select';
 import { useForm } from 'react-hook-form';
@@ -10,7 +11,7 @@ import { House, User } from '../../../api';
 import { houseActions, userActions } from '../../../store/actions/index';
 import { OrangeGlobalState } from '../../../store/state';
 
-import './HouseManageModal.css';
+import './HouseManageModal.scss';
 
 interface Props {
   house?: House;
@@ -55,19 +56,88 @@ function HouseManageModal(props: Props) {
   const onSubmitToToss = () => {
     onTossLeader(props.house!.id, selectedOption!.id);
   };
+  const onRenameHouse = (houseId: number, houseName: string) => {
+    dispatch(houseActions.renameHouse(houseId, houseName));
+  };
+  const onReintroduceHouse = (houseId: number, houseIntroduction: string) => {
+    dispatch(houseActions.reintroduceHouse(houseId, houseIntroduction));
+  };
 
   useEffect(() => {
     userActions.getMe();
   }, [getMeStatus]);
 
-  const formTitle = `${props.house?.name} 관리`;
   const TossLeaderIcon = <i className="far fa-handshake fa-2x" />;
   const LeaveHouseIcon = <i className="fas fa-sign-out-alt fa-2x" />;
   const RemoveHouseIcon = <i className="fas fa-trash-alt fa-2x" />;
   return (
     <>
-      <h2>{formTitle}</h2>
+      <div className="house-modal-name-intro">
+        <h2 className="house-modal-name">
+          {props.house!.users.map((user) => (
+            user.username === me.username && user.is_leader)).includes(true)
+            ? (
+              <EdiText
+                viewContainerClassName="house-modal-name-update-box"
+                editButtonContent={<i className="fas fa-pencil-alt" />}
+                saveButtonContent={<i className="fas fa-check" />}
+                cancelButtonContent={<i className="fas fa-times" />}
+                hideIcons
+                type="text"
+                showButtonsOnHover
+                submitOnUnfocus
+                submitOnEnter
+                cancelOnEscape
+                inputProps={{
+                  className: 'house-modal-name-update-input',
+                  placeholder: 'House 이름을 입력하세요.',
+                  style: { fontSize: 20 },
+                }}
+                validationMessage="한 글자 이상, 열 글자 이하로 입력하세요."
+                validation={(val) => (val.length > 0 && val.length <= 10)}
+                value={props.house!.name}
+                onSave={(houseName: string) => {
+                  onRenameHouse(props.house!.id, houseName);
+                }}
+              />
+            )
+            : props.house!.name}
+        </h2>
+        <div className="house-modal-intro">
+          {props.house!.users.map((user) => (
+            user.username === me.username && user.is_leader)).includes(true)
+            ? (
+              <EdiText
+                viewContainerClassName="house-modal-intro-update-box"
+                editButtonContent={<i className="fas fa-pencil-alt" />}
+                saveButtonContent={<i className="fas fa-check" />}
+                cancelButtonContent={<i className="fas fa-times" />}
+                hideIcons
+                type="text"
+                showButtonsOnHover
+                submitOnUnfocus
+                submitOnEnter
+                cancelOnEscape
+                inputProps={{
+                  className: 'house-modal-intro-update-input',
+                  placeholder: 'House 소개를 입력하세요.',
+                  style: { fontSize: 15 },
+                }}
+                validationMessage="스무 글자 이하로 입력하세요."
+                validation={(val) => val.length <= 20}
+                value={props.house!.introduction}
+                onSave={(houseIntroduction: string) => {
+                  onReintroduceHouse(props.house!.id, houseIntroduction);
+                }}
+              />
+            )
+            : props.house!.introduction}
+        </div>
+      </div>
       <div className="members-box">
+        <h4 className="house-modal-text">
+          집 멤버
+        </h4>
         {props.house!.users?.map((user) => (
           <div
             key={user.id}
@@ -85,14 +155,15 @@ function HouseManageModal(props: Props) {
         ))}
       </div>
       {props.house?.users.map(
-          (user) => (user.username === me.username && user.is_leader)).includes(true)
+        (user) => (user.username === me.username && user.is_leader),
+      ).includes(true)
         ? (
           <>
             <form
               onSubmit={handleSubmit(onSubmitToToss)}
               className="house-toss-leader-form"
             >
-              <h4>
+              <h4 className="house-modal-text">
                 Leader 넘기기&emsp;
               </h4>
               <div className="house-user-container">
